+++ HISTORY.rst
@@ -16,6 +16,7 @@ Release: TBD
 * Remove context manager API from Connection.  It was for transaction.
   New context manager API for closing connection will be added in future version.
 
+* Remove ``waiter`` option from Connection.
 
 ======================
  What's new in 1.3.14